fix setup for vscode
Previously, the `compile_commands.json` file was only searched in the
output directory of the `default` CMake configurePreset
(`/workdir/out/build/default`). That meant that, if you used the
`clang-tidy` preset (without building the `default` preset first), it
wouldn't find it, and the clangd LSP server would scream at you.

Now we tell CMake to copy that file to `/workdir/out/build/` regardless
of the preset, and point there so that it always finds it.